requirements.txt
asgiref==3.3.1
Django==3.1.7
django-sass==1.0.1
libsass==0.20.1
mysql-connector-python==8.0.23
protobuf==3.15.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
